chainntfns: add option to send all confirmations
In this commit, we add a new option for the existing confirmation
notification system that optionally allows the caller to specify that
they wish to receive a notification for every confirmation of the
target transaction.

Since users can now receive transaction details for each confirmation
through the confirmation channel, this commit also removes the Update
channel and updates/adds the corresponding tests to reflect this change.

Signed-off-by: Nishant Bansal <nishant.bansal.282003@gmail.com>